refactor: split out middleware
